Use standard method instead of function invocation to quell prototype warnings
Using correction suggested here:
http://tech.groups.yahoo.com/group/perl-beginner/message/20413

Wed Jan 26, 2005 2:24 pm
Re: [PBML] no warnings (was: no strict)

From: Paul Archer <tigger@...>
> OK, I was tired when I was looking at my code. The errors I'm getting
> are actually warnings, not fatal errors from 'strict'. But the problem
> remains: I am using a module that throws a warning. I don't want to
> modify the module, but I want the warning to go away. Doing: no
> warnings; use Text::Query::ParseAdvanced; use warnings;
>
> doesn't help.

Maybe it's the time to tell us what message do you get exactly and
give us a short self-contained snippet of code that exhibits the
problem.

Looking at the code of the module I do think the authors should drop
the prototypes completely, plus they should never call the methods
like this:
factor($self)
but instead like this:
$self->factor()

If someone tried to subclass the Text::Query::ParseAdvanced class
this would make a big difference.

The
factor($self)
might be a teeny tiny bit quicker, but that should not matter. The
extendability of the code is more important.

Jenda
===== Jenda@Krynicky.cz === http://Jenda.Krynicky.cz =====
When it comes to wine, women and song, wizards are allowed
to get drunk and croon as much as they like.
-- Terry Pratchett in Sourcery